requirements.txt
Appium-Python-Client>=4.0.0
pytest>=8.1.1
PyYAML>=6.0.1
black>=24.3.0
